Added support for binary data
The RedisTimeSeries API is cool, but makes a strong presumption on the
fact that the timestamped data is of type double, that allows simple
maths on aggregation functions. All the rest of the API (range
queries, labelling, downsampling, rules ... ) is of interest for
non-scalar data.
This commit adds support for binary data. In can coexist with
scalar data, so that the API remains 100% compatible.
Just like scalar time series, binary time series can be created with
'TS.CREATE' or 'TS.ADD', just by adding the 'BLOB' keyword in the command.

Aggregation is supported but obviously, performing maths on blob
is nonsense, so that only aggregation types of 'count', 'first'
and 'last' are allowed.

Persistence on disk is also supported.

The testsuite has been completed with the test of the new
features, and passes 100%

Signed-off-by: Thierry Bultel <thierry.bultel@iot.bzh>