Upgrade Redis Time Series to microsecond time resolution
The Redis time series has only millisecond time resolution. The timestamps are
expected to be UNIX timestamp in milliseconds. That resolution is not good
enough for high-frequency data, for example trading data.

This change updates the TS. commands so that the commands accepts as timestamp
microseconds instead.

Since the timestamps as numbers by large passing timestamp as microsecond
timestamps just works (64 bit integers used as timestamps are able to hold the
timestamps).

This change just adds a cosmetic changes, so that TS.ADD with timestamp * adds a
timestamp in microseconds, instead of milliseconds.

Also retention policy code has been updated to convert retention to
microseconds.

As art of this change, the Redis SDK was upgraded to version 7.0. This is needed
since method Redis_microseconds() is available only since version 7.0.