RED-130389 Fix potential crash on TS.RANGE / TS.GET with LATEST argument (#1607)
In case of a compaction key, if the source key is deleted and created
back, it may not have a compaction rule for the destination key
anymore. When the destination key is queried with TS.RANGE family
commands or TS.GET with LATEST flag, it hits an assert when the rule is
not found in find_rule() function.

Scenario:
```
ts.create src
ts.create dst
ts.createrule src dst AGGREGATION sum 1
del src
ts.add src 1 1 // Recreates the key.
ts.range dst - + LATEST AGGREGATION sum 10 // assert
```