MOD-5413 Prevent non-positive values for COUNT argument in TS.RANGE (#1610)
Currently, user can specify non-positive value for COUNT argument in 
TS.RANGE family commands.

If it's -1, command is executed as if like COUNT argument is not given at all. 
If it is 0 or any other negative value, empty array will be returned. 

Added validation to reject any non-positive value for COUNT argument.